Fix su_seconds_t handling
su_seconds_t type used in struct timeval was handled as a long int type.
This is not true on Mac OS X standard library.

This patch fixes it by relying only on POSIX requirements (a signed integer
large enough to accept values from [-1, 1000000] range).

Thanks to Boris Dušek for reporting this issue.

+++ client/common.c
@@ -387,7 +387,7 @@ void print_timeval(const struct timeval *time) {
     if (!localtime_r(&(time->tv_sec), &broken)) goto error;
     if (!strftime(buffer, sizeof(buffer)/sizeof(char), "%c", &broken))
         goto error;
-    printf("%s, %ld us\n", buffer, time->tv_usec);
+    printf("%s, %jd us\n", buffer, (intmax_t)time->tv_usec);
     return;
 
 error:
